Make spec name override system cleaner
